+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added make commands `update-dev` and `update-prod` to update development and production environments respectively after new code is pulled. remp/remp#1281
 - Added hermes handler to notify CRM about refreshing user's data. You can enable the feature in `config.local.neon` (see example file for reference). remp/web#2061
 - Fixed typo in the `package.json` definition for `moment-timezone` causing Yarn3 installation issues. [remp2020/mailer-module#2](https://github.com/remp2020/mailer-module/pull/2) 
+- Added new emit of `user-subscribed-variant` and `user-unsubscribed-variant` hermes event when user subscribe or unsubscribes from the mail type variant. remp/web#2061
 
 ### [Sso]